+++ R/metawas.r
@@ -1,3 +1,106 @@
+#' Benchmark mice imputation
+#' 
+#' Hide a percent of the observations and compare
+#' the results of imputation via \link{mice::mice}
+#' with the original observations.
+#'
+#' @param geno A data table in BIMBAM format. Must contain
+#' one row per SNP, the first three columns must be site_id,
+#' minor_allele and major allele, followed by one column per
+#' sample. Missing values must be encoded as NA.
+#' @param snp A data table with snp information in BIMBAM
+#' format. Must contain columns ID, pos and chr in that order.
+#' Column ID must correspond to column site_id in geno.
+#' @param outdir Output directory to write the results.
+#' @param p Fraction of observations to hide.
+#' @param m Number of imputations performed. See \link{mice::mice}
+#' documentation.
+#' @param verbose Whether to print progress on imputation. 
+#' @param seed Seed for random sambling of data and for imputation
+#' if needed.
+#'
+#' @return A list with elements r, p.imputed, res, and imputed_geno_file.
+#' @export
+#' 
+#' @importFrom magrittr %>%
+benchmark_imputation <- function(geno, snp, outdir, p = 0.1 ,m = 5, verbose = FALSE, seed = NA){
+  dir.create(outdir)
+  
+  # Select positions to impute
+  # geno <- midas_bimbam$Dat$geno
+  gen_only <- geno %>% dplyr::select(-site_id, -minor_allele, -major_allele)
+  
+  if (!is.na(seed)){
+    set.seed(seed)
+  }
+  res <- dplyr::as_tibble(which(!is.na(gen_only), arr.ind = TRUE))
+  res <- res %>%
+    dplyr::bind_cols(hide = sample(c(0,1),
+                                   prob = c(1 - p, p),
+                                   size = nrow(.), replace = TRUE)) %>%
+    filter(hide == 1)
+  
+  # Collect observations
+  gen_only <- gen_only %>% as.matrix
+  ii <- res %>%
+    dplyr::select(row, col) %>%
+    as.matrix
+  res$observed <- gen_only[ii]
+  
+  # Hide data
+  gen_only[ii] <- NA
+  geno_hidden <- geno %>%
+    dplyr::select(site_id, minor_allele, major_allele) %>%
+    dplyr::bind_cols(dplyr::as_tibble(gen_only))
+  
+  # Impute
+  t <- system.time(imp <- mice_impute(geno = geno_hidden,
+                                      snp = snp,
+                                      outdir = outdir,
+                                      m = m,
+                                      verbose = verbose,
+                                      prefix = "imputed",
+                                      return_table = TRUE,
+                                      seed = seed))
+  
+  # Check imputation output
+  if( any(imp$imp$site_id != geno_hidden$site_id)){
+    stop("ERROR")
+  }
+  if( any(colnames(imp$imp) != colnames(geno_hidden))){
+    stop("ERROR")
+  }
+  
+  # Collect imputed values
+  res$imputed <- (imp$imp %>%
+                    select(-site_id, -minor_allele, -major_allele) %>%
+                    as.matrix)[ii]
+  res$path <- outdir
+  
+  # Calculate correlation
+  r <- cor(res$observed, res$imputed, use = "complete.obs")
+  p.imputed <- 1 - (is.na(res$imputed) / nrow(res))
+  
+  # Plot
+  p1 <- ggplot(res, aes(x = observed, y = imputed)) +
+    geom_point() +
+    geom_smooth(method = "lm") +
+    AMOR::theme_blackbox()
+  filename <- file.path(outdir, "observed_vs_imputed.svg")
+  ggsave(filename, p1, width = 5, height = 5)
+  
+  p1 <- res %>%
+    gather(key = "Type", value = "allele_frequency", observed, imputed) %>%
+    ggplot(aes(x=allele_frequency)) +
+    facet_grid(Type ~ .) +
+    geom_histogram(bins = 20) +
+    AMOR::theme_blackbox()
+  filename <- file.path(outdir, "alllele_freq_histograms.svg")
+  ggsave(filename, p1, width = 12, height = 5)
+  
+  return(list(r = r, p.imputed = p.imputed, res = res, imputed_geno_file = imp$imputed_file))
+}
+
 #' Impute genotypes with BIMBAM
 #' 
 #' @param geno_file Path to mean genotype file. See BIMBAM docummentation